Extracted and moved jwtAuth service and unauthorized interceptor in core package
(cherry picked from commit 0bd7bd2)
