+++ src/js/angular/namespaces/controllers.js
@@ -1,12 +1,12 @@
 import 'angular/core/services';
-import 'angular/security/services';
+import 'angular/core/services/jwt-auth.service';
 import 'angular/core/services/repositories.service';
 import 'angular/rest/rdf4j.repositories.rest.service';
 
 const modules = [
     'ui.bootstrap',
     'graphdb.framework.core.services.repositories',
-    'graphdb.framework.security.services',
+    'graphdb.framework.core.services.jwtauth',
     'graphdb.framework.rest.repositories.service',
     'graphdb.framework.rest.rdf4j.repositories.service',
     'toastr'
